--- interfaces/FormInterface.ts
@@ -1,10 +1,0 @@
-export default interface FormInterface {
-    name: string;
-    address: string;
-    city: string;
-    state: string;
-    zip: string;
-    phone: string;
-    email: string;
-    website: string;
-}
