+++ interfaces/Institution/Form.ts
@@ -1,0 +1,5 @@
+export default interface FormInstitutite {
+    name: string;
+    address: string;
+    website: string;
+}
