+++ interfaces/volunteer/Form.ts
@@ -1,0 +1,6 @@
+export default interface FormVolunteer {
+    name: string;
+    email: string;
+    phone: string;
+    password: string;
+}
